requirements.txt
# Web API
python-dotenv>=1.0,<2
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1

# Data & persistence
pydantic>=2.6,<3
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
pgvector>=0.2.4,<0.4

# LLM stack
langchain-core>=0.2,<0.4
langchain-openai>=0.1,<0.4
langchain-huggingface>=0.0.3,<0.3
sentence-transformers>=2.2,<4
langgraph>=0.2,<0.4
typing-extensions>=4.9
